requirements.txt
openai-agents==0.0.9
python-dotenv
pydantic
dateparser

# Testing tools
pytest